requirements.txt
docker_py==1.10.6
pandas==2.3.1
python-dotenv==1.1.1
Requests==2.32.4
supabase==2.18.1
